+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [UNRELEASED]
+* always lowercase the http method
+
 ## 0.3.1
 
 * added `--url` parameter support